requirements.txt
Django>=4.0
django-cors-headers==3.10.1
djangorestframework==3.13.1
django-request-logging==0.7.3
gunicorn==20.1.0
ipython==7.30.1
python-dotenv==0.19.2
python-jose==3.3.0
PyYAML==6.0
uritemplate==4.1.1
